\name{KataegisPortal}
\alias{KataegisPortal}
\alias{KataegisPortal}
\docType{package}
\title{
\Detection and Visualization of Kataegis
}
\description{
\KataegisPortal is an R package that provides utilities to detect and visualize the Kataegis events from single nucleotide variants (SNVs).
}
\section{Main functions}{

\itemize{ 
  \item \code{\link{mutSNP.input}} 
  \item \code{\link{katPoint}} 
  \item \code{\link{mutDis.plot}} 
  \item \code{\link{baseSpe.plot}}
  }
}
\author{
\Mei-Chun Cai <caimeichun@hotmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

