\name{KataegisPortal}
\alias{KataegisPortal}
\alias{KataegisPortal}
\docType{package}
\title{
\packageTitle{KataegisPortal}
}
\description{
\packageDescription{KataegisPortal}
}
\section{Main functions}{

\itemize{ 
  \item \code{\link{mutSNP.input}} 
  \item \code{\link{katPoint}} 
  \item \code{\link{mutDis.plot}} 
  \item \code{\link{baseSpe.plot}}
  }
}
\author{
\packageAuthor{KataegisPortal}
}
\references{
~~ Literature or other references for background information ~~
}
}
